Tweak on questions retrieval
Now the question type needs to be filtered when retrieveing the
questions. Also, the alreadyShown field
is not taken into account. However the question retrieved is random
using the sample capabilities of Mongo.
The above implementation will only work for the regular question modes
and not for the custom ones.
Scheduling-service now retrieves 100 questions per type as some errors
appear.