Changed how questions are fetched
The first implementation used the fetch-custom-question data. the
downside of this is that if some question type failed, the ones that
followed would not be even tried to be fetched.
Now the questions are fetched iterating over the types. This way we
always at least try to get the questions for all the types.