Modified retrieval of questions for custom mode
I tried to reuse as much as the exisiting code as possible. This lead to
a bit of abusive use of local storage of the browser. However it seems
to be functional.

The tests will need to be reviewed as the implementation of some parts
has changed quite a bit.